更新整体

+++ src/main/java/com/itsv/itsvdashboard/dto/CleanDataDto.java
@@ -110,7 +110,7 @@ public class CleanDataDto {
     /**
      * 处理时长
      */
-    private Long processingTime;
+    private Double processingTime;
 
     /**
      * 日期类型
