+++ src/main/java/com/itsv/itsvdashboard/dto/SlaDto.java
@@ -50,7 +50,7 @@ public class SlaDto {
     /**
      * 处理时长
      */
-    private Long processingTime;
+    private Double processingTime;
 
     /**
      * 日期类型
